requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
Flask-Login>=0.7
